refactor(types): remove dead trajectory_store, relocate compute_trajectory_positions
The Trajectory dataclass, its TrajectoryStore alias, and TrajectoryBuilder in unirl/types/trajectory_store.py were entirely unused — the LatentSegment SoA container superseded them. Only the pure helper compute_trajectory_positions remained live, imported by 11 model diffusion stages and the sglang rollout track builder.

- Move compute_trajectory_positions into unirl/types/sampling.py, beside is_forward_process / total_samples_per_prompt
- Repoint all 12 import sites; delete trajectory_store.py
- Drop the dead TrajectoryStore lazy export from unirl.types
- Fix the orphaned docstring reference in segments/latent.py
