+++ unirl/models/bagel/diffusion.py
@@ -48,9 +48,8 @@ from unirl.config.require import require
 from unirl.models.types.diffusion import DiffusionStage
 from unirl.models.types.replay_result import ReplayResult
 from unirl.sde.kernels import FlowSDEStrategy, StepStrategy
-from unirl.types.sampling import DiffusionSamplingParams
+from unirl.types.sampling import DiffusionSamplingParams, compute_trajectory_positions
 from unirl.types.segments.latent import LatentSegment
-from unirl.types.trajectory_store import compute_trajectory_positions
 from unirl.utils.dtypes import parse_torch_dtype
 
 from . import rl_ops
