+++ unirl/models/flux2_klein/diffusion.py
@@ -58,8 +58,8 @@ import torch
 from unirl.models.types.diffusion import DiffusionStage, DiffusionStep
 from unirl.models.types.replay_result import ReplayResult
 from unirl.sde.kernels import StepStrategy
+from unirl.types.sampling import compute_trajectory_positions
 from unirl.types.segments.latent import LatentSegment
-from unirl.types.trajectory_store import compute_trajectory_positions
 from unirl.utils.dtypes import parse_torch_dtype
 
 from .bundle import Flux2KleinBundle
